+++ src/main/java/com/biblioteca/biblioteca/controller/BookController.java
@@ -3,11 +3,8 @@ package com.biblioteca.biblioteca.controller;
 import com.biblioteca.biblioteca.dtos.request.BookRequestDTO;
 import com.biblioteca.biblioteca.dtos.response.BookResponseDTO;
 import com.biblioteca.biblioteca.entities.Book;
-import com.biblioteca.biblioteca.entities.Loan;
+import com.biblioteca.biblioteca.exception.NotFoundException;
 import com.biblioteca.biblioteca.service.BookService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,17 +17,22 @@ import java.util.Optional;
 @RequestMapping("/book")
 public class BookController {
 
-    @Autowired
     private BookService service;
+    public BookController(BookService service) {
+        this.service = service;
+    }
 
     @GetMapping
-    public List<BookResponseDTO> bookList(){
-        return service.getAllBooks();
+    public ResponseEntity<List<BookResponseDTO>> bookList(){
+        List<BookResponseDTO> books = service.getAllBooks();
+        return ResponseEntity.ok(books);
     }
 
     @GetMapping("/{id}")
-    public Optional<BookResponseDTO> bookId(@PathVariable Long id){
-        return service.getById(id);
+    public ResponseEntity<BookResponseDTO> bookId(@PathVariable Long id){
+        return service.getById(id)
+                .map(ResponseEntity::ok)
+                .orElseThrow(() -> new NotFoundException(String.format("O livro com o id %d não foi encontrado", id)));
     }
 
     @PostMapping
@@ -42,7 +44,13 @@ public class BookController {
 
     @DeleteMapping("/{id}")
     public void deleteBook(@PathVariable Long id){
-        service.deleteBook(id);
+        Optional<BookResponseDTO> book  = service.getById(id);
+        if (book.isPresent()){
+            service.deleteBook(id);
+        } else {
+          throw new NotFoundException(String.format("O livro de id %d não foi encontrado", id));
+        }
+
     }
 
 }
